IoTCloudServices/microservices/routes_microservice/code/requirements.txt
paho-mqtt
requests
flask
flask-cors
mysql-connector-python
